пр 11

+++ src/utils/api.js
@@ -46,20 +46,20 @@
         .then(this._checkResponse)
     }
 
-    setLike(cardId) {
-        return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-            method: 'PUT',
-            headers: this._headers,
-        })
-        .then(this._checkResponse);
-    }
-
-    deleteLike(cardId) {
-        return fetch(this._baseUrl + `/cards/${cardId}/likes`, {
-            method: 'DELETE',
-            headers: this._headers,
-        })
-        .then(this._checkResponse);
+    changeLikesCardStatus(cardId, isLiked) {
+        if(!isLiked) {
+            return fetch(this._baseUrl + `/cards/likes/${cardId}`, {
+                method: 'PUT',
+                headers: this._headers,
+            })
+            .then(this._checkResponse)
+        } else {
+            return fetch(this._baseUrl + `/cards/likes/${cardId}`,{
+                method: 'DELETE',
+                headers: this._headers,
+            })
+            .then(this._checkResponse)
+        }
     }
 
     deleteCard(cardId) {
@@ -78,7 +78,14 @@
         })
         .then(this._checkResponse);
     }
+
+
+
+
 }
+
+ 
+
 
 export const api = new Api({
     baseUrl: 'https://mesto.nomoreparties.co/v1/cohort-76',
